+++ backend/controllers/post.js
@@ -26,7 +26,8 @@ exports.createPost = (req, res) => {
                     lastname: userInfos.lastname,
                     avatar: userInfos.avatar
                 });
-                db.query(`INSERT INTO posts(title, content, userId, firstname, lastname, avatar)VALUES(?, ?, ?, ?, ?, ?)`,
+                db.query(`INSERT INTO posts(title, content, userId, firstname, lastname, avatar)VALUES(?, ?, ?, ?, ?, ?
+                          )`,
                     [post.title, post.content, post.userId, post.firstname, post.lastname, post.avatar])
                 res.status(201).json({
                     message: 'Article pubié !',
@@ -42,7 +43,7 @@ exports.createPost = (req, res) => {
 exports.getOnePost = (req, res) => {
     try {
         const postId = req.params;
-        db.query(`SELECT * FROM posts WHERE postId = ?`, postId, (err, row) => {
+        db.query(`SELECT * FROM posts WHERE postId = ?`, postId, (err, row) => { // récupération d'un article avec son id
             if (err || row.length === 0) { // si aucun résultat ou erreur
                 res.status(401).json({message: 'Article non trouvé !'})
             } else { // si article trouvé
@@ -75,14 +76,33 @@ exports.getAllPosts = (req, res) => {
 // fonction de suppression d'un article
 exports.deletePost = (req, res) => {
     try {
-        db.query(`SELECT * FROM posts WHERE postId = ?`, req.query.postId, (err, row) => {
+        const token = req.headers.authorization.split(' ')[1]; // récupération du token dans le header authorization
+        const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET'); // vérification du token pour authentification de l'utilisateur
+        const userId = decodedToken.userId;
+
+        db.query(`SELECT * FROM posts WHERE postId = ?`, req.query.postId, (err, row) => { // récupération d'un article avec son id
             if (err || row.length === 0) { // si aucun résultat ou erreur
-                console.log(1)
-                res.status(401).json({message: 'Impossible de supprimer l article !'})
+                res.status(401).json({message: 'Article non trouvé !'})
             } else { // si article trouvé
-                console.log(2)
-                db.query(`DELETE FROM posts WHERE postId = ?`, req.query.postId); // suppression de l'article sélectionné
-                res.status(200).json({message: 'Article supprimé !'})
+                const postUserId = row[0].userId;
+                if (postUserId !== userId) { // si l'utilisateur connecté n'est pas l'auteur de l'article en question
+                    db.query(`SELECT admin FROM users WHERE userId = ?`, userId, (err, row) => { // recherche du statut de l'utilisateur connecté
+                        if (err || row.length === 0) { // si aucun résultat ou erreur
+                            res.status(401).json({message: 'Informations non trouvées !'})
+                        } else {
+                            const admin = row[0].admin;
+                            if (admin === 1) { // si l'utilisateur connecté est le modérateur
+                                db.query(`DELETE FROM posts WHERE postId = ?`, req.query.postId); // suppression de l'article sélectionné
+                                res.status(200).json({message: 'Article supprimé !'})
+                            } else { // si l'utilisateur connecté n'est pas le modérateur
+                                res.status(401).json({message: 'Impossible de supprimer l article !'})
+                            }
+                        }
+                    })
+                } else { // si l'utilisateur connecté est l'auteur de l'article en question
+                    db.query(`DELETE FROM posts WHERE postId = ?`, req.query.postId); // suppression de l'article sélectionné
+                    res.status(200).json({message: 'Article supprimé !'})
+                }
             }
         })
     } catch (error) {
@@ -108,18 +128,23 @@ exports.likePost = (req, res) => {
                 db.query(`SELECT * FROM likes WHERE userId = ? AND postId =
                           ?`, [userId, req.query.postId], (err, row) => { // recherche d'un like de l'utilisateur connecté pour l'article sélectionné
                     if (row.length === 0) { // si pas de like
-                        db.query(`SELECT * FROM dislikes WHERE userId = ? AND postId = ?`, [userId, req.query.postId], (err, row) => { // recherche d'un dislike de l'utilisateur connecté pour l'article sélectionné
+                        db.query(`SELECT * FROM dislikes WHERE userId = ? AND postId =
+                                  ?`, [userId, req.query.postId], (err, row) => { // recherche d'un dislike de l'utilisateur connecté pour l'article sélectionné
                             if (row.length === 0) { // si pas de dislike
                                 db.query(`INSERT INTO likes(userId, postId)VALUES(?, ?)`, [userId, req.query.postId]); // sauvegarde du like de l'utilisateur pour l'article sélectionné
-                                db.query(`UPDATE posts SET posts.likes = ? WHERE postId = ?`, [likes, req.query.postId]); // ajout du like au total des likes pour l'article sélectionné
+                                db.query(`UPDATE posts SET posts.likes = ? WHERE postId =
+                                          ?`, [likes, req.query.postId]); // ajout du like au total des likes pour l'article sélectionné
                                 res.status(200).json({message: 'L utilisateur aime cet article !'});
                             } else if (err) { // si erreur
                                 res.status(401).json({err})
                             } else { // si dislike
-                                db.query(`DELETE FROM dislikes WHERE postId = ? AND userId = ?`, [req.query.postId, userId]); // suppression du dislike de l'utilisateur pour l'article sélectionné
-                                db.query(`UPDATE posts SET posts.dislikes = ? WHERE postId = ?`, [undislikes, req.query.postId]); // déduction du dislike au total des dislikes pour l'article sélectionné
+                                db.query(`DELETE FROM dislikes WHERE postId = ? AND userId =
+                                          ?`, [req.query.postId, userId]); // suppression du dislike de l'utilisateur pour l'article sélectionné
+                                db.query(`UPDATE posts SET posts.dislikes = ? WHERE postId =
+                                          ?`, [undislikes, req.query.postId]); // déduction du dislike au total des dislikes pour l'article sélectionné
                                 db.query(`INSERT INTO likes(userId, postId)VALUES(?, ?)`, [userId, req.query.postId]); // sauvegarde du like de l'utilisateur pour l'article sélectionné
-                                db.query(`UPDATE posts SET posts.likes = ? WHERE postId = ?`, [likes, req.query.postId]);// ajout du like au total des likes pour l'article sélectionné
+                                db.query(`UPDATE posts SET posts.likes = ? WHERE postId =
+                                          ?`, [likes, req.query.postId]);// ajout du like au total des likes pour l'article sélectionné
                                 res.status(200).json({message: 'L utilisateur aime cet article !'});
                             }
                         })
@@ -145,7 +170,8 @@ exports.getPostUserLike = (req, res) => {
         const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET'); // vérification du token pour authentification  de l'utilisateur
         const userId = decodedToken.userId;
 
-        db.query(`SELECT likes.userId FROM likes WHERE userId = ? AND postId = ?`, [userId, req.query.postId], (err, row) => { // recherche d'un like de l'utilisateur connecté pour l'article sélectionné
+        db.query(`SELECT likes.userId FROM likes WHERE userId = ? AND postId =
+                  ?`, [userId, req.query.postId], (err, row) => { // recherche d'un like de l'utilisateur connecté pour l'article sélectionné
             if (row.length === 0) { // si pas de like
                 console.log("L'utilisateur n'a pas liké cet article");
             } else if (err) { // si erreur
@@ -175,29 +201,38 @@ exports.dislikePost = (req, res) => {
             if (err || row.length === 0) { // si aucun résultat ou erreur
                 return res.status(401).json({err})
             } else { // si article trouvé
-                    db.query(`SELECT * FROM dislikes WHERE userId = ? AND postId = ?`, [userId, req.query.postId], (err, row) => { // recherche d'un dislike de l'utilisateur connecté pour l'article sélectionné
-                        if (row.length === 0) { // si pas de dislike
-                            db.query(`SELECT * FROM likes WHERE userId = ? AND postId = ?`, [userId, req.query.postId], (err, row) => { // recherche d'un like de l'utilisateur connecté pour l'article sélectionné
-                                if (row.length === 0) { // si pas de like
-                                    db.query(`INSERT INTO dislikes(userId, postId) VALUES (?, ?)`, [userId, req.query.postId]); // sauvegarde du dislike de l'utilisateur pour l'article sélectionné
-                                    db.query(`UPDATE posts SET posts.dislikes = ? WHERE postId = ?`, [dislikes, req.query.postId]); // ajout du dislike au total des dislikes pour l'article sélectionné
-                                    return res.status(200).json({message: 'L utilisateur n aime pas cet article !'});
+                db.query(`SELECT * FROM dislikes WHERE userId = ? AND postId =
+                          ?`, [userId, req.query.postId], (err, row) => { // recherche d'un dislike de l'utilisateur connecté pour l'article sélectionné
+                    if (row.length === 0) { // si pas de dislike
+                        db.query(`SELECT * FROM likes WHERE userId = ? AND postId =
+                                  ?`, [userId, req.query.postId], (err, row) => { // recherche d'un like de l'utilisateur connecté pour l'article sélectionné
+                            if (row.length === 0) { // si pas de like
+                                db.query(`INSERT INTO dislikes(userId, postId)VALUES(?, ?
+                                          )`, [userId, req.query.postId]); // sauvegarde du dislike de l'utilisateur pour l'article sélectionné
+                                db.query(`UPDATE posts SET posts.dislikes = ? WHERE postId =
+                                          ?`, [dislikes, req.query.postId]); // ajout du dislike au total des dislikes pour l'article sélectionné
+                                return res.status(200).json({message: 'L utilisateur n aime pas cet article !'});
                             } else if (err) { // si erreur
                                 return res.status(401).json({err})
                             } else { // si like
-                                    db.query(`DELETE FROM likes WHERE postId = ? AND userId = ?`, [req.query.postId, userId]); // suppression du like de l'utilisateur pour l'article sélectionné
-                                    db.query(`UPDATE posts SET posts.likes = ? WHERE postId = ?`, [unlikes, req.query.postId]); // déduction du like du total des likes pour l'article sélectionné
-                                    db.query(`INSERT INTO dislikes(userId, postId) VALUES (?, ?)`, [userId, req.query.postId]); // sauvegarde du dislike de l'utilisateur pour l'article sélectionné
-                                    db.query(`UPDATE posts SET posts.dislikes = ? WHERE postId = ?`, [dislikes, req.query.postId]); // ajout du dislike au total des dislikes pour l'article sélectionné
-                                    return res.status(200).json({message: 'L utilisateur n aime pas cet article !'});
+                                db.query(`DELETE FROM likes WHERE postId = ? AND userId =
+                                          ?`, [req.query.postId, userId]); // suppression du like de l'utilisateur pour l'article sélectionné
+                                db.query(`UPDATE posts SET posts.likes = ? WHERE postId =
+                                          ?`, [unlikes, req.query.postId]); // déduction du like du total des likes pour l'article sélectionné
+                                db.query(`INSERT INTO dislikes(userId, postId)VALUES(?, ?
+                                          )`, [userId, req.query.postId]); // sauvegarde du dislike de l'utilisateur pour l'article sélectionné
+                                db.query(`UPDATE posts SET posts.dislikes = ? WHERE postId =
+                                          ?`, [dislikes, req.query.postId]); // ajout du dislike au total des dislikes pour l'article sélectionné
+                                return res.status(200).json({message: 'L utilisateur n aime pas cet article !'});
                             }
                         })
                     } else if (err) { // si erreur
                         return res.status(401).json({err})
                     } else { // si dislike
-                            db.query(`DELETE FROM dislikes WHERE postId = ? AND userId = ?`, [req.query.postId, userId]); // suppression du dislike de l'utilisateur pour l'article sélectionné
-                            db.query(`UPDATE posts SET posts.dislikes = ? WHERE postId = ?`, [undislikes, req.query.postId]); // déduction du dislike du totat des dislikes pour l'article sélectionné
-                            return res.status(200).json({message: 'dislike supprimé !'})
+                        db.query(`DELETE FROM dislikes WHERE postId = ? AND userId = ?`, [req.query.postId, userId]); // suppression du dislike de l'utilisateur pour l'article sélectionné
+                        db.query(`UPDATE posts SET posts.dislikes = ? WHERE postId =
+                                  ?`, [undislikes, req.query.postId]); // déduction du dislike du totat des dislikes pour l'article sélectionné
+                        return res.status(200).json({message: 'dislike supprimé !'})
                     }
                 })
             }
@@ -213,8 +248,9 @@ exports.getPostUserDislike = (req, res) => {
         const token = req.headers.authorization.split(' ')[1]; // récupération du token dans le header authorization
         const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET'); // vérification du token pour authentification  de l'utilisateur
         const userId = decodedToken.userId;
-         db.query(`SELECT dislikes.userId FROM dislikes WHERE userId = ? AND postId = ?`, [userId, req.query.postId], (err, row) => { // recherche d'un dislike de l'utilisateur connecté pour l'article sélectionné
-                if (row.length === 0) { // si pas de dislike
+        db.query(`SELECT dislikes.userId FROM dislikes WHERE userId = ? AND postId =
+                  ?`, [userId, req.query.postId], (err, row) => { // recherche d'un dislike de l'utilisateur connecté pour l'article sélectionné
+            if (row.length === 0) { // si pas de dislike
                 console.log("L'utilisateur n'a pas disliké cet article")
             } else if (err) { // si erreur
                 return res.status(401).json({err})
